% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FastCooc.R
\docType{data}
\name{coocs-r6}
\alias{coocs-r6}
\alias{Coocc}
\title{Fast co-occurrences calculation}
\format{R6Class}
\usage{
x <- Coocc$new(binDTM)
x$ccoocs
}
\arguments{
\item{binDTM}{a DocumentTermMatrix in the form of a Matrix::dgCMatrix}

\item{minCoocFreq}{a number specifying the minimal times a cooccurrence has to appear, for it not beeing set to 0}

\item{maxCoocFreq}{a number specifying the maximal times a cooccurrence can appear, for it not beeing set to 0}

\item{significanceThreshhold}{a number which determindes a threshold. Significances below this threshold will be set to zero.}

\item{measure}{The significance measure (DICE,LOGLIK,MI,COUNT). Defaults to DICE.}
}
\value{
a significance-matrix based on the given \code{binDTM}
}
\description{
Calculating co-occurrences of binary document-term-matrices
}
\section{Methods}{

  \describe{
    \item{\code{ccoocs()}}{Calculates co-occurences of binary DocumentTermMatrices and returns a significance matrix.}
    \item{\code{set_binDTM(binDTM)}}{Set the binary DocumentTermMatrix.}
    \item{\code{set_minCoocFreq(minCoocFreq)}}{Set the minCoocFreq.}
    \item{\code{set_maxCoocFreq(maxCoocFreq)}}{Set the maxCoocFreq.}
    \item{\code{set_significanceThreshold(significanceThreshold)}}{Set the significanceThreshold.}
    \item{\code{set_measure(measure)}}{Set the measure.}
    \item{\code{get_binDTM()}}{Returns the binary DocumentTermMatrix.}
    \item{\code{get_minCoocFreq()}}{Returns the minCoocFreq.}
    \item{\code{get_maxCoocFreq()}}{Returns the maxCoocFreq.}
    \item{\code{get_significanceThreshold()}}{Returns the significanceThreshold.}
    \item{\code{get_measure()}}{Returns the measure.}
    }
}

\keyword{datasets}
